melo/api.py: add a 'tts' iterator to greatly improve the response speed
In 'tts_to_file' function, the preprocessing process will try to split
a long sentences into texts array. Then using model to interfence with
each sentences and combine the result into finally audio array. But if
the sentences is very very long, wait the entire process to be finished
will cost lots of time, it is not a great idea. Most of the time, the
interfence speed will extremely faster than playing speed, so use a
iterator to get each of the audio piece.